% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_plate.R
\name{coord_plate}
\alias{coord_plate}
\title{Plate coordinates}
\usage{
coord_plate(
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  default = FALSE,
  clip = "off",
  corner = "right",
  specs = 96
)
}
\arguments{
\item{xlim}{Limits for the x and y axes.}

\item{ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{default}{Is this the default coordinate system? If \code{FALSE} (the default),
then replacing this coordinate system with another one creates a message alerting
the user that the coordinate system is being replaced. If \code{TRUE}, that warning
is suppressed.}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. In most cases, the default of \code{"on"} should not be changed,
as setting \code{clip = "off"} can cause unexpected results. It allows
drawing of data points anywhere on the plot, including in the plot margins. If
limits are set via \code{xlim} and \code{ylim} and some data points fall outside those
limits, then those data points may show up in places such as the axes, the
legend, the plot title, or the plot margins.}

\item{corner}{The place to draw the notched corner of a plate. Can be one of
\code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"topleft"},
\code{"topright"}, \code{"bottomleft"}, \code{"bottomright"}, \code{"none"} or \code{"all"}.}

\item{specs}{Either a \code{list} with plate dimensions or a \code{numeric(1)} with
one of the common formats, e.g. \code{6}, \code{12}, \code{24}, \code{48}, \code{96} or \code{384} wells.}
}
\value{
A \code{Coord} object that can be added to a plot.
}
\description{
This coordinate functions makes your plot look more like a microwell plate.
}
\details{
The \code{panel.background} theme element now draws a polygon instead
of a rectangle in case \code{corner != "none"}. Moreover, the \code{panel.grid.major}
element is used to draw circles at well positions and \code{panel.grid.minor} is
used to draw regular grid lines. Actual minor breaks are ignored.
}
\examples{
NULL
}
