% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplateplot-package.R
\docType{package}
\name{ggplateplot-package}
\alias{ggplateplot-package}
\alias{_PACKAGE}
\title{ggplateplot: Plots for plates in cell culture}
\description{
Make plots for data from cell culture plates, such as 96 well plates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teunbrand/ggplateplot}
  \item Report bugs at \url{https://github.com/teunbrand/ggplateplot/issues}
}

}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
