% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_plates.R
\name{new_plate_spec}
\alias{new_plate_spec}
\alias{custom_plate_spec}
\title{Define a new plate layout}
\usage{
new_plate_spec(
  width = 127.8,
  height = 85.5,
  hor_spacing = NULL,
  ver_spacing = NULL,
  well_spacing = 9,
  well_diameter = 6.8,
  corner_size = 8,
  ncol = 12L,
  nrow = 8L
)

custom_plate_spec(ncol = 12L, nrow = 8L, width = 127.8, height = 85.5)
}
\arguments{
\item{width}{Width of the plate in millimetres.}

\item{height}{Height of the plate in millimetres.}

\item{hor_spacing}{Horizontal spacing between wells in millimetres. If
\code{NULL} (default), the \code{well_spacing} argument will be used instead.}

\item{ver_spacing}{Vertical spacing between wells in millimetres. If \code{NULL}
(default), the \code{well_spacing} argument will be used instead.}

\item{well_spacing}{Spacing between wells in millimetres. Useful when
horizontal and vertical spacing is identical.}

\item{well_diameter}{Diameter of a well in millimetres.}

\item{corner_size}{Size of the adjacent/opposite sides of the triangle, in
millimeters, when plate has a corner cut off.}

\item{ncol}{The number of wells in horizontal direction.}

\item{nrow}{The number of wells in vertical direction.}
}
\value{
A \code{list} with dimensions that can be understood by the \code{specs}
argument in \code{coord_plate()}.
}
\description{
The \code{new_plate_spec()} function is used to define the specification of a
plate layout. The \code{custom_plate_spec()} is a wrapper around
\code{new_plate_spec()} that heuristically sets dimensions for quick and dirty
layouts.
}
\examples{
# Precisely detail a new layout
rotated_96well <- new_plate_spec(
  width         = 85.5,
  height        = 127.8,
  well_spacing  = 9,
  well_diameter = 6.8,
  ncol          = 8,
  nrow          = 12
)

## Setup dummy data
df <- expand.grid(row = 1:12, col = LETTERS[1:8])
df$value <- seq_len(nrow(df))

## Use new spec in plot
ggplot(df, aes(col, row, fill = value)) +
  geom_well() +
  coord_plate(spec = rotated_96well)

# Quickly sketch a new layout
square_2x2 <- custom_plate_spec(ncol = 2, nrow = 2, width = 100, height = 100)

## Setup dummy data
df <- expand.grid(row = 1:2, col = LETTERS[1:2])
df$value <- seq_len(nrow(df))

## Use spec in plot
ggplot(df, aes(col, row, fill = value)) +
  geom_well() +
  coord_plate(spec = square_2x2)
}
